OccupantItem links to profile
