+++ src/longbeach/location/encampment/OccupantList.js
@@ -1,33 +1,20 @@
 // @flow
 import React from 'react';
 
-import styled from 'styled-components';
 import { faUserSlash } from '@fortawesome/pro-solid-svg-icons';
 import {
   Card,
   CardSegment,
   IconSplash,
-  MinusButton,
   Spinner,
 } from 'lattice-ui-kit';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { RequestStates } from 'redux-reqseq';
 
-import { removePersonFromEncampment } from './EncampmentActions';
+import OccupantItem from './OccupantItem';
 import { ENCAMPMENT_STORE_PATH } from './constants';
 
-import { getFirstLastFromPerson } from '../../../utils/PersonUtils';
-
-const StyledSegment = styled(CardSegment)`
-  padding: 10px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex: 1 0 auto;
-`;
-
 const OccupantList = () => {
-  const dispatch = useDispatch();
   const livesAt = useSelector((store) => store.getIn([...ENCAMPMENT_STORE_PATH, 'occupation', 'livesAt']));
   const people = useSelector((store) => store.getIn([...ENCAMPMENT_STORE_PATH, 'occupation', 'people']));
   const fetchState = useSelector((store) => store.getIn([...ENCAMPMENT_STORE_PATH, 'occupation', 'fetchState']));
@@ -53,15 +40,7 @@ const OccupantList = () => {
       {
         livesAt.toJS().map((edge) => {
           const person = people.get(edge);
-          const name = getFirstLastFromPerson(person);
-
-          const deleteEdge = () => dispatch(removePersonFromEncampment(edge));
-          return (
-            <StyledSegment key={edge} padding="10px">
-              <span>{name}</span>
-              <MinusButton size="sm" mode="negative" onClick={deleteEdge} />
-            </StyledSegment>
-          );
+          return <OccupantItem key={edge} person={person} livesAtEKID={edge} />;
         })
       }
     </Card>
